+++ CHANGELOG.md
@@ -25,6 +25,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Make the exception codes 32-bit safe (#287)
 - Remove associated subscriptions when a subscriber list or subscriber is removed (#271, #272)
 - Always truncate the DB tables after an integration test (#259)
 - Adapt the system tests to possible HTTP errors (#256)